requirements.txt
# Input Output Realated Librarys
anyio
asgiref
atomicwrites
attrs
bcrypt
certifi
cffi
charset-normalizer
coloredlogs
dnspython
ecdsa
email-validator
fastapi
greenlet
h11
httptools
humanfriendly
idna
iniconfig
itsdangerous
Jinja2
MarkupSafe
orjson
packaging
passlib
pluggy
py
pyasn1
pycparser
pydantic
PyJWT
pyparsing
pyreadline3
pytest
python-dotenv
python-jose
python-multipart
pytz
PyYAML
requests
rsa
six
sniffio
SQLAlchemy
SQLAlchemy-Utils
starlette
toml
twilio
typing-extensions
ujson
urllib3
uvicorn
watchgod
fastapi_mail
PyMuPDF
psycopg2
alembic
